Add -Werror and update code. Add dependencies. Rename TenantServer to Main. Refactor packages to better fit real implementation.
