+++ build.sbt
@@ -24,5 +24,5 @@ lazy val tenant = project.in(file("tenant"))
   .configure(C.akkaPersistentEntity("improving-app-tenant"))
   .dependsOn(commonTypes % "compile->compile;test->test")
 
-(Compile / runMain) := ((Compile / runMain) in tenant).evaluated
+(Compile / runMain) := (tenant / Compile / runMain).evaluated
 onLoad in Global := (onLoad in Global).value andThen (Command.process("project tenant", _))
