+++ tenant/src/main/protobuf/com/improving/app/tenant/tenantService.proto
@@ -1,9 +1,9 @@
 syntax = "proto3";
 
-package com.tenant;
+package com.improving.app.tenant;
 
 option java_multiple_files = true;
-option java_package = "com.tenant";
+option java_package = "com.improving.app.tenant";
 option java_outer_classname = "TenantServiceProto";
 
 service TenantService {
