+++ tenant/src/main/protobuf/com/improving/app/tenant/testMessages.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package com.tenant;
+package com.improving.app.tenant;
 
 message TestCommand {
   string myInput = 1;
